Fix directories structure
