+++ app/go.mod
@@ -1,3 +1,5 @@
 module app
 
 go 1.21
+
+require github.com/gorilla/mux v1.8.1 // indirect
