+++ app/pkg/config/app.go
@@ -1,0 +1,22 @@
+package config
+
+import (
+	"github.com/jinzhu/gorm"
+	_ "github.com/jinzhu/gorm/dialects/mysql"
+)
+
+var (
+	db * gorm.DB
+)
+
+func Connect(){
+	d, err: = gorm.Open("mysql", "albertolionardi:1234/ SOMETHING ")
+	if err != nil{
+		panic(err)
+	}
+	db = d
+}
+
+func GetDB() *gorm.DB{
+	return db
+}
